Separate out memory benchmarks in flowbench
