+++ flowbench/BenchMem.cpp
@@ -1,0 +1,48 @@
+/*
+ * BenchMem.cpp
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2020 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "benchmark/benchmark.h"
+
+static void bench_memcmp(benchmark::State& state) {
+	constexpr int kLength = 10000;
+	std::unique_ptr<char[]> b1{ new char[kLength] };
+	std::unique_ptr<char[]> b2{ new char[kLength] };
+	memset(b1.get(), 0, kLength);
+	memset(b2.get(), 0, kLength);
+	b2.get()[kLength - 1] = 1;
+
+	while (state.KeepRunning()) {
+		benchmark::DoNotOptimize(memcmp(b1.get(), b2.get(), kLength));
+	}
+}
+
+static void bench_memcpy(benchmark::State& state) {
+	constexpr int kLength = 10000;
+	std::unique_ptr<char[]> b1{ new char[kLength] };
+	std::unique_ptr<char[]> b2{ new char[kLength] };
+	memset(b1.get(), 0, kLength);
+
+	while (state.KeepRunning()) {
+		benchmark::DoNotOptimize(memcpy(b2.get(), b1.get(), kLength));
+	}
+}
+
+BENCHMARK(bench_memcmp);
+BENCHMARK(bench_memcpy);
